+++ airbyte-integrations/connectors/source-mongodb-internal-poc/.dockerignore
