+++ airbyte-integrations/connectors/source-mongodb-internal-poc/Dockerfile
@@ -10,7 +10,7 @@ FROM airbyte/integration-base-java:dev AS build
 
 WORKDIR /airbyte
 
-ENV APPLICATION source-mongodb-v3
+ENV APPLICATION source-mongodb-internal-poc
 
 COPY build/distributions/${APPLICATION}*.tar ${APPLICATION}.tar
 
@@ -20,9 +20,9 @@ FROM airbyte/integration-base-java:dev
 
 WORKDIR /airbyte
 
-ENV APPLICATION source-mongodb-v3
+ENV APPLICATION source-mongodb-internal-poc
 
 COPY --from=build /airbyte /airbyte
 
 LABEL io.airbyte.version=0.0.1
-LABEL io.airbyte.name=airbyte/source-mongodb-v3
+LABEL io.airbyte.name=airbyte/source-mongodb-internal-poc
