+++ airbyte-integrations/connectors/source-mongodb-internal-poc/acceptance-test-config.yml
@@ -1,6 +1,6 @@
 # See [Connector Acceptance Tests](https://docs.airbyte.com/connector-development/testing-connectors/connector-acceptance-tests-reference)
 # for more information about how to configure these tests
-connector_image: airbyte/source-mongodb-v2:dev
+connector_image: airbyte/source-mongodb-internal-poc:dev
 acceptance_tests:
   spec:
     tests:
