+++ airbyte-integrations/connectors/source-mongodb-internal-poc/acceptance-test-docker.sh
