+++ airbyte-integrations/connectors/source-mongodb-internal-poc/integration_tests/acceptance.py
