+++ airbyte-integrations/connectors/source-mongodb-internal-poc/metadata.yaml
@@ -3,8 +3,8 @@ data:
   connectorType: source
   definitionId: 5ac5a7e5-43f5-4e7a-bf53-70961b0307bc
   dockerImageTag: 0.0.1
-  dockerRepository: airbyte/source-mongodb-v3
-  githubIssueLabel: source-mongodb-v3
+  dockerRepository: airbyte/source-mongodb-internal-poc
+  githubIssueLabel: source-mongodb-internal-poc
   icon: mongodb.svg
   license: ELv2
   name: MongoDb
@@ -14,7 +14,7 @@ data:
     oss:
       enabled: true
   releaseStage: alpha
-  documentationUrl: https://docs.airbyte.com/integrations/sources/mongodb-v3
+  documentationUrl: https://docs.airbyte.com/integrations/sources/mongodb-internal-poc
   tags:
     - language:java
 metadataSpecVersion: "1.0"
