+++ airbyte-integrations/connectors/source-mongodb-internal-poc/src/main/resources/spec.json
@@ -1,6 +1,6 @@
 {
-  "documentationUrl": "https://docs.airbyte.com/integrations/sources/mongodb-v3",
-  "changelogUrl": "https://docs.airbyte.com/integrations/sources/mongodb-v3",
+  "documentationUrl": "https://docs.airbyte.com/integrations/sources/mongodb-internal-poc",
+  "changelogUrl": "https://docs.airbyte.com/integrations/sources/mongodb-internal-poc",
   "connectionSpecification": {
     "$schema": "http://json-schema.org/draft-07/schema#",
     "title": "MongoDb Source Spec",
